requirements.txt
Django==1.10.2
psycopg2
biopython
decorator==4.0.2
django-guardian
django-crispy-forms
django-extensions
djangorestframework
django-filter
django-taggit
django-taggit-templatetags
django-form-utils
python-social-auth==0.3.5
requests
requests-oauthlib
opbeat
ua-parser
user-agents
django-user-agents
django-dirtyfields
cython
python-openid
django-sendfile
django-polymorphic
celery[redis]==3.1.25
django-celery
django-cleanup
git+https://github.com/sinnwerkstatt/django-file-resubmit.git#egg=file_resubmit
django-hstore==1.3.5
django-datatables-view
django-oauth-toolkit==0.9.0
django-rest-swagger
django-gravatar2
simplejson
pygments
numpy
social-auth-app-django
django-notifications-hq
som
sendgrid
python3-saml
